[NFC] Add observability subfolder
