+++ .github/workflows/pr-test.yml
@@ -595,7 +595,15 @@ jobs:
             },
             {
               "num_gpus": 0,
-              "test_file": "test_train_dump.py"
+              "test_file": "test_train_data_utils.py"
+            },
+            {
+              "num_gpus": 0,
+              "test_file": "test_rollout_data_utils.py"
+            },
+            {
+              "num_gpus": 0,
+              "test_file": "test_rollout_metrics.py"
             },
             {
               "num_gpus": 0,
@@ -616,6 +624,10 @@ jobs:
             {
               "num_gpus": 0,
               "test_file": "test_logprob_response_spans.py"
+            },
+            {
+              "num_gpus": 0,
+              "test_file": "observability/test_trace_utils.py"
             },
             {
               "num_gpus": 0,
