+++ slime/backends/megatron_utils/data.py
@@ -1,28 +1,14 @@
-import logging
-from argparse import Namespace
 from collections.abc import Sequence
 
-import numpy as np
 import torch
-import torch.distributed as dist
 import torch.nn.functional as F
 from megatron.core import mpu
 from megatron.core.packed_seq_params import PackedSeqParams
 
-from slime.utils import accelerator, train_metric_utils
-from slime.utils.flops_utils import calculate_fwd_flops
-from slime.utils.metric_utils import compute_pass_rate, compute_rollout_step
+from slime.utils import accelerator
 from slime.utils.types import RolloutBatch
 
-from ...utils import logging_utils
-from .cp_utils import (
-    gather_and_reduce_log_dict,
-    get_sum_of_sample_mean,
-    rollout_log_metric_contribution,
-    slice_with_cp,
-)
-
-logger = logging.getLogger(__name__)
+from .cp_utils import slice_with_cp
 
 
 def get_batch(
@@ -163,41 +149,6 @@ def get_batch(
     return batch
 
 
-def gather_log_data(
-    metric_name: str,
-    args: Namespace,
-    rollout_id: int,
-    log_dict: dict[str, "float | tuple[float, float]"],
-) -> dict[str, float] | None:
-    """
-    Gather per-rank metrics, reduce on the DP source rank, and log to W&B / TB.
-
-    Each value in ``log_dict`` is either:
-      * a ``(sum, count)`` tuple → reduced as ``Σsum / Σcount``;
-      * a plain scalar → reduced as ``Σ / dp_size`` (mean across ranks).
-
-    The gather + reduce step is delegated to
-    :func:`cp_utils.gather_and_reduce_log_dict` so it can be exercised by
-    CPU multi-process unit tests directly. This function adds the
-    ``metric_name`` prefix and the W&B / TB logging side effects.
-    """
-    reduced = gather_and_reduce_log_dict(
-        log_dict,
-        dp_size=mpu.get_data_parallel_world_size(with_context_parallel=True),
-        dp_src_rank=mpu.get_data_parallel_src_rank(with_context_parallel=True),
-        dp_group=mpu.get_data_parallel_group_gloo(with_context_parallel=True),
-    )
-    if reduced is None:
-        return None
-    reduced_log_dict = {f"{metric_name}/{k}": v for k, v in reduced.items()}
-    logger.info(f"{metric_name} {rollout_id}: {reduced_log_dict}")
-    # Calculate step once to avoid duplication
-    step = compute_rollout_step(args, rollout_id)
-    reduced_log_dict["rollout/step"] = step
-    logging_utils.log(args, reduced_log_dict, step_key="rollout/step")
-    return reduced_log_dict
-
-
 class DataIterator:
     """Iterator over a rollout dict following an explicit micro-batch index schedule."""
 
@@ -243,277 +194,6 @@ def get_data_iterator(rollout_data: RolloutBatch) -> list[DataIterator]:
     vpp_size = mpu.get_virtual_pipeline_model_parallel_world_size() or 1
     micro_batch_indices = rollout_data["micro_batch_indices"]
     return [DataIterator(rollout_data, micro_batch_indices) for _ in range(vpp_size)]
-
-
-def log_rollout_data(
-    rollout_id: int,
-    args: Namespace,
-    rollout_data: RolloutBatch,
-) -> None:
-    """
-    Summarize rollout fields and log reduced metrics on PP last stage, TP rank 0.
-
-    - Tensor-valued lists are concatenated and averaged. For token-level metrics
-      like log-probs/returns/advantages/values, computes a CP-correct sample mean
-      using `loss_masks` and total/response lengths.
-    - Non-tensor lists are averaged elementwise.
-    - Scalars are converted to Python numbers.
-    """
-    if mpu.get_tensor_model_parallel_rank() == 0 and mpu.is_pipeline_last_stage():
-        cp_size = mpu.get_context_parallel_world_size()
-        log_dict = {}
-        response_lengths = rollout_data["response_lengths"]
-        loss_masks = rollout_data["loss_masks"]
-        total_lengths = rollout_data["total_lengths"]
-        # Same per-rollout denominators the training loss uses, so reported
-        # log_probs / returns / advantages / etc. live in the same per-rollout
-        # mean space (rather than per-sample) as the gradient signal.
-        rollout_mask_sums = rollout_data.get("rollout_mask_sums", None)
-        # For per-rollout-mean metrics: ``rollout_log_metric_contribution``
-        # produces the ``(sum, count)`` tuple so gather_log_data's
-        # ``Σsum / Σcount`` lands on ``sum_DP_full / num_rollouts`` — the
-        # same number train_one_step reports for the same samples.
-        dp_world = mpu.get_data_parallel_world_size(with_context_parallel=False)
-        num_rollouts_in_rollout = sum(rollout_data["global_batch_sizes"])
-
-        for key, val in rollout_data.items():
-            if key in [
-                "tokens",
-                "multimodal_train_inputs",
-                "loss_masks",
-                "sample_indices",
-                "rollout_ids",
-                "rollout_mask_sums",
-                "rollout_top_p_token_ids",
-                "rollout_top_p_token_offsets",
-                "rollout_routed_experts",
-                "global_batch_sizes",
-                "num_microbatches",
-                "micro_batch_indices",
-                "source_names",
-                # DP-local view of `raw_reward`, which this loop already logs;
-                # both reduce to the same mean, so skip the duplicate metric.
-                "local_raw_reward",
-            ]:
-                continue
-            # Emit (sum, count) so gather_log_data can do a weighted average across
-            # DP ranks. This stops the legacy "every rank has the same N samples"
-            # assumption from biasing means once uneven-DP partitioning lands.
-            if isinstance(val, (list, tuple)):
-                count = len(val)
-                if isinstance(val[0], torch.Tensor):
-                    # NOTE: Here we have to do the clone().detach(), otherwise the tensor will be
-                    # modified in place and will cause problem for the next rollout.
-                    if key in [
-                        "log_probs",
-                        "ref_log_probs",
-                        "rollout_log_probs",
-                        "returns",
-                        "advantages",
-                        "values",
-                        "teacher_log_probs",
-                        "opd_reverse_kl",
-                    ]:
-                        tensor = torch.cat(val).clone().detach()
-                        sum_of_sample_mean = get_sum_of_sample_mean(
-                            total_lengths,
-                            response_lengths,
-                            loss_masks,
-                            rollout_mask_sums,
-                        )
-                        # Compute (sum, count) via the shared helper so this
-                        # path and the unit tests stay in sync.
-                        sum_value, count = rollout_log_metric_contribution(
-                            sum_of_sample_mean(tensor).item(),
-                            cp_size=cp_size,
-                            num_rollouts_in_rollout=num_rollouts_in_rollout,
-                            dp_size=dp_world,
-                        )
-                        log_dict[key] = (sum_value, count)
-                        continue
-                    tensor = torch.cat(val).clone().detach()
-                    # val.mean() * cp_size is the per-sample mean for one rank;
-                    # multiply by count to get the per-rank sum.
-                    per_rank_sum = tensor.mean() * cp_size * count
-                    sum_value = per_rank_sum.item()
-                else:
-                    sum_value = sum(val)
-                log_dict[key] = (sum_value, count)
-            elif isinstance(val, torch.Tensor):
-                # Scalar tensor (one per rank): treat as count=1.
-                log_dict[key] = (val.float().mean().item(), 1)
-            else:
-                raise ValueError(f"Unsupported type: {type(val)} for key: {key}")
-
-        reduced_log_dict = gather_log_data("rollout", args, rollout_id, log_dict)
-        if args.ci_test and reduced_log_dict is not None:
-            # This is an initial actor/ref zero-KL check. R3 replays rollout
-            # routing for the actor forward, while the reference forward
-            # intentionally falls through to normal routing, so their
-            # log-probs are not expected to match bit-for-bit in CI.
-            if (
-                rollout_id == 0
-                and not getattr(args, "ci_disable_kl_checker", False)
-                and not getattr(args, "use_rollout_routing_replay", False)
-                and "rollout/log_probs" in reduced_log_dict
-                and "rollout/ref_log_probs" in reduced_log_dict
-            ):
-                # TODO: figure out why there is a small numerical difference in log_probs and ref_log_probs in CI test, and whether it's expected or not.
-                # assert reduced_log_dict["rollout/log_probs"] == reduced_log_dict["rollout/ref_log_probs"]
-                assert abs(reduced_log_dict["rollout/log_probs"] - reduced_log_dict["rollout/ref_log_probs"]) < 1e-8
-            if "rollout/log_probs" in reduced_log_dict:
-                assert -1 < reduced_log_dict["rollout/log_probs"] < 0
-            if "rollout/entropy" in reduced_log_dict:
-                assert 0 < reduced_log_dict["rollout/entropy"] < 1
-
-    if args.log_multi_turn:
-        log_multi_turn_data(rollout_id, args, rollout_data)
-    if args.log_passrate:
-        log_passrate(rollout_id, args, rollout_data)
-
-    if args.log_correct_samples:
-        if mpu.get_tensor_model_parallel_rank() == 0 and mpu.is_pipeline_last_stage():
-            cp_size = mpu.get_context_parallel_world_size()
-            log_dict = {}
-            response_lengths = rollout_data["response_lengths"]
-            loss_masks = rollout_data["loss_masks"]
-            total_lengths = rollout_data["total_lengths"]
-
-            def quantile(total_value, n_quantiles, data) -> dict:
-                import math
-
-                assert n_quantiles > 1, f"n_quantiles({n_quantiles}) must be greater than 1."
-
-                quantiles = [((i + 1) / n_quantiles) for i in range(n_quantiles)]
-                cut_points = [total_value * q for q in quantiles]
-                cut_points[-1] = total_value
-
-                count = [0] * n_quantiles
-                for d in data:
-                    for i, point in enumerate(cut_points):
-                        if d <= point:
-                            count[i] += 1
-                            break
-
-                total = sum(count) + 1e-9
-                percentile = [c / total for c in count]
-
-                percentile = {f"p{min(math.ceil(q*100),100)}": p for q, p in zip(quantiles, percentile, strict=True)}
-                return percentile
-
-            # DP-local, so it lines up positionally with response_lengths /
-            # total_lengths / loss_masks / log_probs below. `raw_reward` itself
-            # is the whole rollout batch (log_passrate needs the full grouping).
-            raw_rewards = rollout_data["local_raw_reward"]
-            # Additional metrics for correct cases are calculated separately below.
-            correct_response_lengths = []
-            correct_total_lengths = []
-            correct_loss_masks = []
-            correct_entropy = []
-            for i, raw_reward in enumerate(raw_rewards):
-                if raw_reward == 1:
-                    correct_response_lengths.append(response_lengths[i])
-                    correct_total_lengths.append(total_lengths[i])
-                    correct_loss_masks.append(loss_masks[i])
-                    correct_entropy.append(-rollout_data["log_probs"][i])
-            num_correct_responses = len(correct_total_lengths)
-            rollout_data["correct_response_lengths"] = correct_response_lengths
-            correct_response_length_percentile = quantile(
-                args.rollout_max_response_len, 4, rollout_data["correct_response_lengths"]
-            )
-            for p, val in correct_response_length_percentile.items():
-                rollout_data[f"correct_length/{p}"] = [val] * num_correct_responses
-            if len(correct_entropy) > 0:
-                # NOTE: per-sample-mean over the correct subset, not per-rollout.
-                # A rollout's siblings may not all be correct, and slicing
-                # ``rollout_mask_sums`` here would leave a denom that still
-                # includes incorrect siblings — meaningless for a "correct-only"
-                # entropy report. Per-sample-mean over the filtered subset is
-                # the cleanest semantic.
-                sum_of_sample_mean = get_sum_of_sample_mean(
-                    correct_total_lengths, correct_response_lengths, correct_loss_masks, sample_denoms=None
-                )
-                correct_entropy = sum_of_sample_mean(torch.cat(correct_entropy, dim=0))
-                rollout_data["correct_entropy"] = [correct_entropy.item()] * num_correct_responses
-            else:
-                rollout_data["correct_entropy"] = [0] * num_correct_responses
-
-
-def log_multi_turn_data(rollout_id: int, args: Namespace, rollout_data: RolloutBatch) -> None:
-    """
-    Log multi-turn auxiliary metrics such as raw/observed response lengths and rounds.
-
-    Operates only on PP last stage and TP rank 0. Uses GPU tensors when available
-    to compute statistics without host transfers.
-    """
-    if mpu.get_tensor_model_parallel_rank() == 0 and mpu.is_pipeline_last_stage():
-        log_dict = {}
-        for key, val in rollout_data.items():
-            if key == "loss_masks":
-                if val:  # Check if val is not empty
-                    device = val[0].device  # Get device from first tensor
-
-                    # Vectorized length calculation using torch
-                    raw_response_lengths = torch.tensor([v.shape[0] for v in val], dtype=torch.float32, device=device)
-                    log_dict["raw_response_length/response_length_mean"] = raw_response_lengths.mean().item()
-                    log_dict["raw_response_length/response_length_max"] = raw_response_lengths.max().item()
-                    log_dict["raw_response_length/response_length_min"] = raw_response_lengths.min().item()
-                    log_dict["raw_response_length/response_length_clip_ratio"] = (
-                        (raw_response_lengths >= args.rollout_max_response_len).float().mean().item()
-                    )
-
-                    # Vectorized sum calculation using torch - stay on GPU
-                    wo_obs_response_lengths = torch.tensor(
-                        [v.sum().item() for v in val], dtype=torch.float32, device=device
-                    )
-                    log_dict["wo_obs_response_length/response_length_mean"] = wo_obs_response_lengths.mean().item()
-                    log_dict["wo_obs_response_length/response_length_max"] = wo_obs_response_lengths.max().item()
-                    log_dict["wo_obs_response_length/response_length_min"] = wo_obs_response_lengths.min().item()
-            if key == "round_number":
-                # Use numpy for vectorized round number statistics
-                round_number_array = np.array(val)
-                log_dict["multi_turn_metric/round_number_mean"] = np.mean(round_number_array)
-                log_dict["multi_turn_metric/round_number_max"] = np.max(round_number_array)
-                log_dict["multi_turn_metric/round_number_min"] = np.min(round_number_array)
-        gather_log_data("multi_turn", args, rollout_id, log_dict)
-
-
-def log_passrate(rollout_id: int, args: Namespace, rollout_data: RolloutBatch) -> None:
-    """
-    Compute pass@k metrics from `raw_reward` groups and log the results.
-
-    `raw_reward` is reshaped to `[group_number, group_size]`, then pass@k is
-    estimated per problem and averaged.
-    """
-    if mpu.get_tensor_model_parallel_rank() == 0 and mpu.is_pipeline_last_stage():
-        log_dict = {}
-        for key, val in rollout_data.items():
-            if key != "raw_reward":
-                continue
-
-            log_dict |= compute_pass_rate(
-                flat_rewards=val,
-                group_size=args.n_samples_per_prompt,
-                num_groups=args.rollout_batch_size,
-            )
-
-        gather_log_data("passrate", args, rollout_id, log_dict)
-
-
-def log_perf_data(rollout_id: int, args: Namespace, extra_metrics: dict | None = None) -> None:
-    train_metric_utils.log_perf_data_raw(
-        rollout_id=rollout_id,
-        args=args,
-        is_primary_rank=(
-            mpu.get_tensor_model_parallel_rank() == 0
-            and mpu.is_pipeline_last_stage()
-            and mpu.get_data_parallel_rank(with_context_parallel=True) == 0
-        ),
-        compute_total_fwd_flops=lambda seq_lens: calculate_fwd_flops(seqlens=seq_lens, args=args)
-        / dist.get_world_size()
-        / 1e12,
-        extra_metrics=extra_metrics,
-    )
 
 
 def tensors_to_cpu(tensor_list):
