+++ slime/observability/__init__.py
@@ -1,0 +1,1 @@
+"""Training and rollout observability utilities."""
