+++ slime/observability/rollout_metrics.py
@@ -1,0 +1,271 @@
+import logging
+from typing import Any
+
+import numpy as np
+import torch
+
+from slime.observability import logging_utils
+from slime.observability.metric_utils import (
+    compute_pass_rate,
+    compute_rollout_step,
+    compute_statistics,
+    dict_add_prefix,
+    has_repetition,
+)
+from slime.utils.misc import group_by, load_function
+from slime.utils.types import Sample
+
+logger = logging.getLogger(__name__)
+
+_SGLANG_REQUEST_PERF_FIELDS = (
+    ("request/e2e_latency", "e2e_latency"),
+    ("request/queue_time", "queue_time"),
+    ("decode/throughput", "decode_throughput"),
+)
+_SGLANG_PREFILL_PERF_FIELDS = (
+    ("prefill/bootstrap_queue_duration", "pd_prefill_bootstrap_queue_duration"),
+    ("prefill/bootstrap_duration", "pd_prefill_bootstrap_duration"),
+    ("prefill/alloc_wait_duration", "pd_prefill_alloc_wait_duration"),
+    ("prefill/forward_duration", "pd_prefill_forward_duration"),
+    ("prefill/transfer_queue_duration", "pd_prefill_transfer_queue_duration"),
+    ("prefill/transfer_speed_gb_s", "pd_transfer_speed_gb_s"),
+    ("prefill/transfer_total_mb", "pd_transfer_total_mb"),
+    ("prefill/retry_count", "pd_prefill_retry_count"),
+)
+_SGLANG_DECODE_PERF_FIELDS = (
+    ("decode/prealloc_duration", "pd_decode_prealloc_duration"),
+    ("decode/bootstrap_duration", "pd_decode_bootstrap_duration"),
+    ("decode/alloc_wait_duration", "pd_decode_alloc_wait_duration"),
+    ("decode/transfer_duration", "pd_decode_transfer_duration"),
+    ("decode/forward_duration", "pd_decode_forward_duration"),
+)
+
+
+def compute_metrics_from_samples(args, samples):
+    response_lengths = [sample.effective_response_length for sample in samples]
+
+    log_dict = {}
+    log_dict |= dict_add_prefix(compute_statistics(response_lengths), "response_len/")
+    log_dict |= _compute_zero_std_metrics(args, samples)
+    log_dict |= _compute_spec_metrics(args, samples)
+    log_dict |= _compute_prefix_cache_metrics(args, samples)
+    log_dict |= _compute_reward_cat_metrics(args, samples)
+    log_dict |= _compute_top_p_kept_vocab_metrics(args, samples)
+    log_dict["repetition_frac"] = np.mean([int(has_repetition(s.response)) for s in samples]).item()
+    log_dict["truncated_ratio"] = np.mean([int(s.status == Sample.Status.TRUNCATED) for s in samples]).item()
+    return log_dict
+
+
+def compute_perf_metrics_from_samples(args, samples, rollout_time):
+    non_generation_time = [sample.non_generation_time for sample in samples]
+
+    log_dict = {}
+    log_dict["rollout_time"] = rollout_time
+    if max(non_generation_time) > 0:
+        log_dict |= dict_add_prefix(compute_statistics(non_generation_time), "non_generation_time/")
+
+    def token_perf(response_lengths, non_generation_time, key=""):
+        max_response_length = max(response_lengths)
+        if args.rollout_num_gpus:
+            log_dict[f"{key}tokens_per_gpu_per_sec"] = sum(response_lengths) / rollout_time / args.rollout_num_gpus
+        log_dict[f"longest_{key}sample_tokens_per_sec"] = max_response_length / rollout_time
+
+        if max(non_generation_time) == 0:
+            return
+
+        non_generation_time = [
+            t for t, length in zip(non_generation_time, response_lengths, strict=True) if length == max_response_length
+        ]
+        mean_non_generation_time = sum(non_generation_time) / len(non_generation_time)
+
+        log_dict[f"longest_{key}sample_non_generation_time"] = mean_non_generation_time
+        log_dict[f"longest_{key}sample_tokens_per_sec_without_non_generation"] = max_response_length / (
+            rollout_time - mean_non_generation_time
+        )
+
+    token_perf([sample.response_length for sample in samples], non_generation_time, key="")
+    token_perf([sample.effective_response_length for sample in samples], non_generation_time, key="effective_")
+    log_dict |= _compute_sglang_request_perf_metrics(samples)
+
+    return log_dict
+
+
+def _compute_sglang_request_perf_metrics(all_samples: list[Sample]):
+    attrs_by_request = list(_iter_sglang_generate_attrs(all_samples))
+    if not attrs_by_request:
+        return {}
+
+    values_by_metric: dict[str, list[float]] = {}
+    profiled_request_count = 0
+
+    def add_value(metric_key: str, source_key: str, attrs: dict) -> bool:
+        value = attrs.get(source_key)
+        if not isinstance(value, (int, float)) or isinstance(value, bool) or not np.isfinite(value):
+            return False
+        values_by_metric.setdefault(metric_key, []).append(float(value))
+        return True
+
+    for attrs in attrs_by_request:
+        request_has_perf = False
+
+        for metric_key, source_key in _SGLANG_REQUEST_PERF_FIELDS:
+            request_has_perf |= add_value(metric_key, source_key, attrs)
+
+        for metric_key, source_key in _SGLANG_PREFILL_PERF_FIELDS:
+            request_has_perf |= add_value(metric_key, source_key, attrs)
+
+        for metric_key, source_key in _SGLANG_DECODE_PERF_FIELDS:
+            request_has_perf |= add_value(metric_key, source_key, attrs)
+
+        if request_has_perf:
+            profiled_request_count += 1
+
+    metrics: dict[str, float] = {}
+    for key, values in values_by_metric.items():
+        if not values:
+            continue
+        metrics |= dict_add_prefix(compute_statistics(values), f"{key}/")
+
+    return metrics
+
+
+def _iter_sglang_generate_attrs(all_samples: list[Sample]):
+    for sample in all_samples:
+        trace = getattr(sample, "trace", None)
+        if not isinstance(trace, dict):
+            continue
+        for event in trace.get("events") or []:
+            if event.get("type") != "span_end" or event.get("name") != "sglang_generate":
+                continue
+            attrs = event.get("attrs")
+            if isinstance(attrs, dict):
+                yield attrs
+
+
+def _compute_zero_std_metrics(args, all_samples: list[Sample]):
+    # only compute in GRPO-like algorithms where one prompt has multiple responses
+    if args.advantage_estimator == "ppo":
+        return {}
+
+    def _is_zero_std(samples: list[Sample]):
+        rewards = [sample.get_reward_value(args) for sample in samples]
+        return len(rewards) == 0 or all(rewards[0] == r for r in rewards)
+
+    all_sample_groups = group_by(all_samples, lambda s: s.group_index)
+    interesting_sample_groups = [g for g in all_sample_groups.values() if _is_zero_std(g)]
+
+    interesting_rewards = [str(round(g[0].get_reward_value(args), 1)) for g in interesting_sample_groups]
+
+    return {f"zero_std/count_{reward}": len(items) for reward, items in group_by(interesting_rewards).items()}
+
+
+def _compute_top_p_kept_vocab_metrics(args, all_samples: list[Sample]):
+    total_kept = 0
+    total_tokens = 0
+    for sample in all_samples:
+        offsets = sample.rollout_top_p_token_offsets
+        if offsets is None or sample.response_length == 0:
+            continue
+        offsets = torch.as_tensor(offsets, dtype=torch.int64)
+        if offsets.numel() == 0:
+            continue
+        assert (
+            offsets.numel() == sample.response_length + 1
+        ), f"top-p token offsets length {offsets.numel()} != response length + 1 {sample.response_length + 1}"
+        if sample.remove_sample:
+            continue
+        if sample.loss_mask is None:
+            total_kept += int(offsets[-1] - offsets[0])
+            total_tokens += sample.response_length
+            continue
+        loss_mask = torch.as_tensor(sample.loss_mask, dtype=torch.bool, device=offsets.device)
+        assert (
+            loss_mask.numel() == sample.response_length
+        ), f"loss mask length {loss_mask.numel()} != response length {sample.response_length}"
+        total_kept += int(torch.diff(offsets)[loss_mask].sum())
+        total_tokens += int(loss_mask.sum())
+    if total_tokens == 0:
+        return {}
+    return {"top_p_kept_vocab_per_token": total_kept / total_tokens}
+
+
+def _compute_spec_metrics(args, all_samples: list[Sample]):
+    if getattr(args, "sglang_speculative_algorithm", None) is None:
+        return {}
+    num_samples = len(all_samples)
+    metrics = {}
+    metrics["spec_accept_rate"] = sum(sample.spec_info.spec_accept_rate for sample in all_samples) / num_samples
+    metrics["spec_accept_length"] = sum(sample.spec_info.spec_accept_length for sample in all_samples) / num_samples
+    return metrics
+
+
+def _compute_prefix_cache_metrics(args, all_samples: list[Sample]):
+    num_samples = len(all_samples)
+    metrics = {}
+    total_cached_tokens = sum(sample.prefix_cache_info.cached_tokens for sample in all_samples)
+    total_prompt_tokens = sum(sample.prefix_cache_info.total_prompt_tokens for sample in all_samples)
+
+    metrics["prefix_cache_hit_rate"] = total_cached_tokens / total_prompt_tokens if total_prompt_tokens > 0 else 0.0
+    metrics["avg_cached_tokens_per_sample"] = total_cached_tokens / num_samples
+    return metrics
+
+
+def _compute_reward_cat_metrics(args, all_samples: list[Sample]):
+    reward_cat_key = args.log_reward_category
+    if reward_cat_key is None:
+        return {}
+
+    samples_of_reward_cat = group_by(all_samples, lambda s: s.reward[reward_cat_key])
+
+    return {f"error_cat/{reward_cat}": len(s) / len(all_samples) for reward_cat, s in samples_of_reward_cat.items()}
+
+
+def log_eval_rollout_data(rollout_id, args, data, extra_metrics: dict[str, Any] | None = None):
+    if args.custom_eval_rollout_log_function_path is not None:
+        custom_log_func = load_function(args.custom_eval_rollout_log_function_path)
+        if custom_log_func(rollout_id, args, data, extra_metrics):
+            return
+
+    log_dict = extra_metrics or {}
+    for key in data.keys():
+        rewards = data[key]["rewards"]
+        log_dict[f"eval/{key}"] = sum(rewards) / len(rewards)
+        if (samples := data[key].get("samples")) is not None:
+            log_dict |= dict_add_prefix(compute_metrics_from_samples(args, samples), f"eval/{key}/")
+        if "truncated" in data[key]:
+            truncated = data[key]["truncated"]
+            log_dict[f"eval/{key}-truncated_ratio"] = sum(truncated) / len(truncated)
+        if args.log_passrate:
+            log_dict |= dict_add_prefix(
+                compute_pass_rate(
+                    flat_rewards=rewards,
+                    group_size=args.n_samples_per_eval_prompt,
+                ),
+                f"eval/{key}-",
+            )
+
+    logger.info(f"eval {rollout_id}: {log_dict}")
+
+    step = compute_rollout_step(args, rollout_id)
+    log_dict["eval/step"] = step
+    logging_utils.log(args, log_dict, step_key="eval/step")
+
+    return log_dict
+
+
+def log_rollout_data(rollout_id, args, samples, rollout_extra_metrics, rollout_time):
+    if args.custom_rollout_log_function_path is not None:
+        custom_log_func = load_function(args.custom_rollout_log_function_path)
+        if custom_log_func(rollout_id, args, samples, rollout_extra_metrics, rollout_time):
+            return
+
+    if args.load_debug_rollout_data:
+        return
+
+    log_dict = {**(rollout_extra_metrics or {})}
+    log_dict |= dict_add_prefix(compute_metrics_from_samples(args, samples), "rollout/")
+    log_dict |= dict_add_prefix(compute_perf_metrics_from_samples(args, samples, rollout_time), "perf/")
+    logger.info(f"perf {rollout_id}: {log_dict}")
+    step = compute_rollout_step(args, rollout_id)
+    log_dict["rollout/step"] = step
+    logging_utils.log(args, log_dict, step_key="rollout/step")
